+++ NEWS.md
@@ -1,13 +1,15 @@
 climatrends 0.1.2 (2020-04-24)
 =========================
 * Export function `get_timeseries()`
-* Enable modular expansion of data sources using API clients by creating specific functions that are called with the argument `source` in all functions
+* Enable future modular expansion of data sources using API clients by creating specific functions that are called with the argument `source` in all functions
+* Enable argument `last.day`, optional to `span`, when indices are retrieved with focus on timeseries
 * Enable data retrieve from large areas in 'nasapower' by clustering the lonlat points with `stats::hclust()`
 * Methods for objects of classes 'sf', 'matrix', 'array' and 'data.frame' (default) in all functions
 * Remove dependency from 'tibble' which was basically to provide a 'cool' printing method
 * Implement a `print()` method for objects that inherits the class 'clima_df'
 * Calls from "nasapower" can be adjusted with argument `pars`, mainly for `temperature()`, `GDD()` and `ETo()`. Details in the function documentation
-* New `temperature()` indices provided
+* New `temperature()` indices provided. Details in the function documentation
+* 
 
 climatrends 0.1.0 (2020-02-20)
 =========================